Adds new failing test for not replacing placeholders that have no key in data_provider
If we replace everything that looks like it is a placeholder, but
doesn't have a key in the data provider it will make it very hard to
debug why a placeholder isn't being filled in. It will simply be gone.

Key should always exist, just put nil value if no value is used.